requirements.txt
requests
python-dotenv
python-telegram-bot
pandas
scikit-learn
xgboost